final? no Javadocs yet

+++ MCO2/src/ComboItem.java
@@ -2,7 +2,7 @@ public class ComboItem extends Item{
     String pMsg;
     String type;
 
-    public ComboItem(String name, double calories, int price, int initQuantity, 
+    public ComboItem(String name, int calories, int price, int initQuantity, 
     String pMsg, String type) {
         super(name, calories, price, initQuantity);
         this.pMsg = "";
